Коммит 8.
Внедрил pressly/goose:
- Разделил database.go на database.go и dbquery.go;
- Удалил migrations.go;
- Создал в /internal/storage/migration/migrations/ .sql файл с нулевой миграцией;
- Удалил флаги и переменные, описывающие миграции;

Теперь автоматически применяется миграция 0 up, которая проверяет на наличие
таблицы и полей, и в случае отсутствия создаёт их средствами goose. Миграция
down и произвольные миграции возможны извне, например так:
goose -dir "internal/storage/migration/migrations" postgres "postgresql://user:password@localhost:5432/gophermart" down-to 0
