+++ internal/server/handlers/authHandler.go
