+++ internal/server/handlers/authHandler_test.go
